reworked expression grammar to make it PEG compatible.  There were too many possible orderings of various FieldAccess/MethodInvocation/ClassInstanceCreationExpression variants that are LR with primary to come up with a single ordering, so I just rewrote the whole expression grammar from scratch to be PEG compatible.  I reused names where possible.  The major change is that Primary now delegates most of the work to ExpressionAtom and parses PostOps to the right to handle the suffixes for individual field accesses, method invocations, etc.  These are no longer left-recursrive, but are instead a heterogenous n-ary operator.
I added annotation handling to gen_parser_helpers to avoid adding a
node for every precedence tier of binary operators, and (@name=...)
syntax to allow specifying a name, and (@postcond=...) to allow
postconditions which makes it much easier to recombine individual
variants from Primary into productions which match a subset of
expressions like LeftHandSide and StatementExpression.

Fixed bug in packrat caching: when MagicDotIdentifier borrowed an
identifier, if the result was cached, a use of the cached result might
leave the identifier in its original position leading to multiple
instances of the same identifier.

Fixed bug: MagicDotIdentifier was not borrowing when the
ContextFreeNames were not in the same variant that has the ".".

Added a test and some wiring in pom.xml so that we can run the parser
against our own source files.