Fix bugs related to super.method() calls and super.field() accesses.
TypingPass was binding super accesses using the sub-type as a starting
point, instead of skipping the sub-type in favor of method and field
declarations in super-types.

TypingPass was also not treating

   super.foo()

as a reference to a method on `this` so was treating it as equivalent
to

   foo()

which could possibly be resolved in an outer class or against a static
import in the case where foo() is not actually overridden in the containing
type.