Fixed problems that result when type inference resolves to unmentionable types.
When an anonymous class constructor appears in a parameterized
method call, a type parameter can bind to the anonymous class name.
If we then try to insert into the AST a TypeArguments node with
the inferred types, we fail because there is no type name
for that class that can be expressed in Java Source.
Fixed this by electing not to inline TypeArguments nodes in this
case.
This will be mostly mooted in the larger use case because when
after flatten type names, there are no anonymous classes, so the
only unmentionable types are intersection types and
intermediate wildcard types.

The above problem manifested as a failure to validate a serialized
AST because of a TypeNameNode that included the Identifier "1".
Teaked parSer generator to check the new BasseLeafNode.isValidValue
method before allowing a setValue call to complete.
Added some machinery to PTrees to ParSer to allow fastMatching
via a method that normally just invokes parse under the hood
but which is specialized into a fast case for PatternMatch
which is what backs leaf nodes.
This required tweaking class Input since we don't want
"b\u0061r" to be misclassified as a valid Identifier even
though it will parse to one after preprocessing is done.

To solve the above problem, declared a deanonymize operator over
types that replaces an anonymous type reference with its most
specific super type.  To do this it became convenient to define a
TypeSpecification.Mapper class that makes it easy to walk a
TypeSpecification and as recursively as needed, substitute parts
of the output.

A related problem came up in enums.  When the output type of
a parameterized method call is unmentionable, we couldn't figure
out how to convert
     foo()
to
     unmentionable.Type.this.<unmentionable.Type>foo()
since only mentionable types can appear at the front of an
explicit `this` reference.
In many contexts, we can actually just use `this`.  TypingPass
now does this.