Resolved ambiguity: MemberInfo.getMemberInfo in FieldDeclaration
FieldDeclarations can declare multiple members:

  private int x, y;

Changed the type of the memberInfo field in the MemberDeclaration
mixin to a list of MemberInfo and declared a subtype
SoleMemberDeclaration which has convenience methods for the common
case.

Added some code to the AST code generator to add a setMemberInfo
check that ensures that SoleMemberDeclarations only have one
member info.